Restructure admin components into one main folder
