+++ lvlup-frontend/src/components/Admin/admin-add-challenge/add-challenge-main.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import HamburgerAdmin from '../admin-common/hamburger';
-import NavBarLoggedIn from '../navbar-loggedin';
+import NavBarLoggedIn from '../../navbar-loggedin';
 import AddChallengeForm from './add-challenge-form';
