+++ lvlup-frontend/src/components/Admin/admin-challenges/admin-challenges-main.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import HamburgerAdmin from '../admin-common/hamburger';
-import NavBarLoggedIn from '../navbar-loggedin';
+import NavBarLoggedIn from '../../navbar-loggedin';
 import ChallengesTable from './challenges-table';
 import ChallengesHeader from './header';
 import AddButton from './button';
