+++ lvlup-frontend/src/components/Admin/admin-config/admin-config-styles.css
@@ -1,0 +1,4 @@
+.admin-config {
+  text-align: center;
+  float: right;
+}
