+++ lvlup-frontend/src/components/Admin/admin-edit-challenge/admin-edit-challenge-styles.css
