+++ lvlup-frontend/src/components/Admin/admin-edit-reward/edit-reward-main.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import HamburgerAdmin from '../admin-common/hamburger';
-import NavBarLoggedIn from '../navbar-loggedin';
+import NavBarLoggedIn from '../../navbar-loggedin';
 import EditRewardForm from './edit-reward-form';
