+++ lvlup-frontend/src/components/Admin/admin-main-view/admin-main.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import HamburgerAdmin from '../admin-common/hamburger';
-import NavBarLoggedIn from '../navbar-loggedin';
+import NavBarLoggedIn from '../../navbar-loggedin';
 import ChallengesHeader from './pending-challenges/header';
 import SubmissionsHeader from './pending-submissions/header';
 import ChallengesTable from './pending-challenges/challenges-table';
