+++ lvlup-frontend/src/components/Admin/admin-main-view/pending-submissions/submissions-table.js
