+++ lvlup-frontend/src/components/Admin/admin-rewards/admin-rewards-main.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import HamburgerAdmin from '../admin-common/hamburger';
-import NavBarLoggedIn from '../navbar-loggedin';
+import NavBarLoggedIn from '../../navbar-loggedin';
 import RewardsTable from './rewards-table';
 import RewardsHeader from './header';
 import AddButton from './button';
