+++ lvlup-frontend/src/components/Admin/admin-rewards/rewards-table.js
